refactor(check-baselines): thin check-baselines into phase pipeline (resolves #2493)
