+++ .agents/scripts/lib/orchestration/check-baselines/phases/evaluate.js
@@ -1,0 +1,80 @@
+/**
+ * evaluate.js — Phase 4 of the check-baselines pipeline (Story #2466).
+ *
+ * Runs the per-kind pipeline: load → floor → compare → tolerance → report.
+ * Extracted from `check-baselines.js` without behavior change.
+ *
+ * @module lib/orchestration/check-baselines/phases/evaluate
+ */
+
+import { checkKernelVersion } from '../../../baselines/kernel.js';
+import * as reader from '../../../baselines/reader.js';
+import { applyTolerance, evaluateCompare, runCompareStage } from './compare.js';
+import { applyFloors, flattenBreaches } from './floors.js';
+
+function loadHeadBaseline(kind, cwd, configPath) {
+  try {
+    return { baseline: reader.load(kind, { cwd, configPath }) };
+  } catch (err) {
+    const message = err?.message ?? String(err);
+    const tag = /schema validation failed/i.test(message) ? 'schema' : 'read';
+    return { schemaError: { tag, message } };
+  }
+}
+
+function buildGateReport({
+  kind,
+  gateBlock,
+  baseline,
+  findings,
+  breaches,
+  compareOutput,
+  cmp,
+}) {
+  const kernel = checkKernelVersion(kind, baseline.kernelVersion);
+  return {
+    kind,
+    enabled: true,
+    kernelMatch: kernel.match,
+    kernelCurrent: kernel.current,
+    kernelBaseline: baseline.kernelVersion,
+    tolerance: gateBlock.tolerance ?? null,
+    floors: gateBlock.floors ?? {},
+    components: findings,
+    breachCount: breaches.length,
+    breaches,
+    regressions: compareOutput.regressions,
+    improvements: compareOutput.improvements,
+    unchanged: compareOutput.unchanged,
+    additions: compareOutput.additions ?? [],
+    regressionCount: compareOutput.regressions.length,
+    baseRef: cmp.baseRef ?? null,
+    generatedAt: baseline.generatedAt,
+  };
+}
+
+export async function evaluateKind({
+  kind,
+  gateBlock,
+  scope,
+  cwd,
+  configPath,
+}) {
+  const headLoad = loadHeadBaseline(kind, cwd, configPath);
+  if (headLoad.schemaError) return { kind, schemaError: headLoad.schemaError };
+  const baseline = headLoad.baseline;
+  const findings = applyFloors(kind, baseline.rollup, gateBlock.floors ?? {});
+  const breaches = flattenBreaches(findings);
+  const cmp = await evaluateCompare({ kind, gateBlock, scope, cwd });
+  const rawCompare = runCompareStage(baseline, cmp);
+  const compareOutput = applyTolerance(rawCompare, gateBlock.tolerance ?? null);
+  return buildGateReport({
+    kind,
+    gateBlock,
+    baseline,
+    findings,
+    breaches,
+    compareOutput,
+    cmp,
+  });
+}
